requirements.txt
asgiref==3.3.4
autopep8==1.5.6
cffi==1.14.5
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
djongo==1.3.4
dnspython==2.1.0
gunicorn==20.1.0
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pymongo==3.11.3
pytz==2021.1
sqlparse==0.2.4
toml==0.10.2
wfastcgi==3.0.0
whitenoise==5.2.0
